requirements.txt
# Dependencias principales
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.26.2
openpyxl>=3.1.2
PyYAML>=6.0.1

# APIs
openai>=1.10.0,<2.0.0
langchain>=0.1.10,<0.2.0
langchain-openai>=0.0.5
tiktoken>=0.5.2

# LLMWhisperer
llmwhisperer-client>=2.3.1
requests>=2.31.0

# Utilidades
pathlib2>=2.3.0
typing-extensions>=4.0.0
